+++ packages/server/src/queue/consumers/kafka/consumers/subscription-v1/handle-payload.ts
@@ -16,9 +16,9 @@ import { OperationError, outcomeError } from "@iguhealth/operation-outcomes";
 import logAuditEvent, {
   MAJOR_FAILURE,
   createAuditEvent,
-} from "../../../../fhir-logging/auditEvents.js";
-import { resolveOperationDefinition } from "../../../../fhir-operation-executors/utilities.js";
-import { IGUHealthServerCTX } from "../../../../fhir-server/types.js";
+} from "../../../../../fhir-logging/auditEvents.js";
+import { resolveOperationDefinition } from "../../../../../fhir-operation-executors/utilities.js";
+import { IGUHealthServerCTX } from "../../../../../fhir-server/types.js";
 
 export async function handleSubscriptionPayload(
   ctx: IGUHealthServerCTX,
