+++ packages/server/src/queue/consumers/kafka/utilities.ts
