+++ packages/server/src/queue/topics/dynamic-topic.ts
@@ -1,4 +1,4 @@
-import { DynamicMessage } from "../provider/interface.js";
+import { DynamicMessage } from "../providers/interface.js";
 import { ITopic } from "./index.js";
 
 export type DynamicTopic = ITopic<DynamicMessage> & {
